* Reorganization of noise_models into inference * Pylint for examples
