+++ .pylintrc
@@ -1,3 +1,4 @@
+
 [MASTER]
 
 # A comma-separated list of package or module names from where C extensions may
@@ -63,7 +64,8 @@ confidence=
 disable=similarities,
         no-self-use,
         fixme,
-        useless-type-doc
+        useless-type-doc,
+        relative-beyond-top-level
 
 # Enable the message, report, category or checker with the given id(s). You can
 # either give multiple identifier separated by comma (,) or put this option
