+++ examples/09_simple_layer_deterministic_pulses.py
@@ -16,6 +16,7 @@ deterministic pulse trains for update.
 Simple network that consist of one analog layer. The network aims to learn
 to sum all the elements from one array.
 """
+# pylint: disable=invalid-name
 
 # Imports from PyTorch.
 from torch import Tensor
