+++ examples/16_mnist_gan.py
@@ -12,12 +12,16 @@
 
 """aihwkit example 16: Train a GAN using Linear Layers to generate MNIST characters.
 
-This experiment creates a Generative Adversarial Network (GAN) -https://arxiv.org/abs/1406.2661-
-that generates characters (numbers) alla MNIST. The experiment uses analog layers to implement
-the network. Also, while training it plots the different characters generated by the generator
-neural network vs. those used to train the discriminator neural network.
-Also, for every epoch, it shows the loss for the generator and discriminator.
+This experiment creates a Generative Adversarial Network (GAN)
+-https://arxiv.org/abs/1406.2661- that generates characters (numbers)
+alla MNIST. The experiment uses analog layers to implement the
+network. Also, while training it plots the different characters
+generated by the generator neural network vs. those used to train the
+discriminator neural network.  Also, for every epoch, it shows the
+loss for the generator and discriminator.
+
 """
+# pylint: disable=invalid-name
 
 import os
 import glob
@@ -34,26 +38,20 @@ from torchvision.utils import make_grid, save_image
 
 import matplotlib.pyplot as plt
 import matplotlib.image as plt_image
-import matplotlib.animation as animation
+from matplotlib import animation
 
 from aihwkit.nn import AnalogLinear, AnalogSequential
 from aihwkit.optim import AnalogSGD
-from aihwkit.simulator.presets import MixedPrecisionPCMPreset
+from aihwkit.simulator.presets import MixedPrecisionEcRamMOPreset
 
 # Select the device model to use in the training.
 
-# * There are a number of presets available in
+# There are a number of presets available in
 # aihwkit.simulator.presets. This is will also determine the analog
 # optimizer used (e.g. mixed precision or full analog update)
 
-# * If `FloatingPointRPUConfig(device=FloatingPointDevice())` then standard
-#   floating point devices will be used
-
-# RPU_CONFIG = FloatingPointRPUConfig()
-# Parallel SGD in analog with constant step device from Gokmen & Vlasov 2016.
-# RPU_CONFIG = GokmenVlasovPreset()
-# Using digital floating point rank-update and transfer to PCM.
-RPU_CONFIG = MixedPrecisionPCMPreset()
+# As an example we use a mixed precision preset using an ECRAM device model
+RPU_CONFIG = MixedPrecisionEcRamMOPreset()
 WS_OMEGA = 0.8
 
 # Set your parameters
@@ -82,6 +80,13 @@ def store_tensor_images(image_tensor, label, current_step, num_images=25, size=(
 
     Given a tensor of images, number of images, and size per image, stores the
     images using a uniform grid.
+
+    Args:
+        image_tensor (Tensor): tensor of images
+        label (str): text label
+        current_step (int): current step number
+        num_images (int): number of images
+        size (Tuple): shape of images
     """
     image_unflat = image_tensor.detach().cpu().view(-1, *size)
     image_grid = make_grid(image_unflat[:num_images], nrow=5)
@@ -111,11 +116,11 @@ def get_generator_block(input_dim, output_dim):
     dimensions.
 
     Args:
-        input_dim: the dimension of the input vector, a scalar
-        output_dim: the dimension of the output vector, a scalar
+        input_dim (int): the dimension of the input vector, a scalar
+        output_dim (int): the dimension of the output vector, a scalar
 
     Returns:
-        a generator neural network layer, with a linear transformation
+        n.Module: a generator neural network layer, with a linear transformation
             followed by a batch normalization and then a relu activation
     """
     return AnalogSequential(
@@ -165,7 +170,7 @@ class Generator(nn.Module):
         Given a noise tensor, returns generated images.
 
         Args:
-            noise: a noise tensor with dimensions (n_samples, z_dim)
+            noise (Tensor): a noise tensor with dimensions (n_samples, z_dim)
 
         Returns:
             Tensor: the generated images.
@@ -180,9 +185,12 @@ def get_noise(n_samples, z_dim, device='cpu'):
     filled with random numbers from the normal distribution.
 
     Args:
-        n_samples: the number of samples to generate, a scalar
-        z_dim: the dimension of the noise vector, a scalar
-        device: the device type
+        n_samples (int): the number of samples to generate, a scalar
+        z_dim (int): the dimension of the noise vector, a scalar
+        device (device): the device type
+
+    Returns:
+        Tensor: random vector
     """
     # NOTE: To use this on GPU with device='cuda', make sure to pass the device
     # argument to the function you use to generate the noise.
@@ -196,8 +204,8 @@ def get_discriminator_block(input_dim, output_dim):
     and output dimensions.
 
     Args:
-        input_dim: the dimension of the input vector, a scalar
-        output_dim: the dimension of the output vector, a scalar
+        input_dim (int): the dimension of the input vector, a scalar
+        output_dim (int): the dimension of the output vector, a scalar
 
     Returns:
         nn.Sequential: a discriminator neural network layer, with a linear transformation
@@ -213,13 +221,13 @@ class Discriminator(nn.Module):
     """Discriminator Class.
 
     Args:
-        im_dim: the dimension of the images, fitted for the dataset used, a scalar
+        im_dim (int): the dimension of the images, fitted for the dataset used, a scalar
             (MNIST images are 28x28 = 784 so that is your default)
-        hidden_dim: the inner dimension, a scalar
+        hidden_dim (int): the inner dimension, a scalar
     """
 
     def __init__(self, im_dim=784, hidden_dim=128):
-        super(Discriminator, self).__init__()
+        super().__init__()
         self.disc = nn.Sequential(
             get_discriminator_block(im_dim, hidden_dim * 4),
             get_discriminator_block(hidden_dim * 4, hidden_dim * 2),
@@ -233,7 +241,7 @@ class Discriminator(nn.Module):
         Given an image tensor, returns a 1-dimension tensor representing fake/real.
 
         Args:
-            image: a flattened image tensor with dimension (im_dim)
+            image (Tensor): a flattened image tensor with dimension (im_dim)
 
         Returns:
             Tensor: a 1-dimension tensor representing fake/real
@@ -245,19 +253,23 @@ def get_disc_loss(gen, disc, criterion, real, num_images, z_dim, device):
     """Return the loss of the discriminator given inputs.
 
     Args:
-        gen: the generator model, which returns an image given z-dimensional noise
-        disc: the discriminator model, which returns a single-dimensional prediction of real/fake
-        criterion: the loss function, which should be used to compare
-            the discriminator's predictions to the ground truth reality of the images
-            (e.g. fake = 0, real = 1)
-        real: a batch of real images
-        num_images: the number of images the generator should produce,
-            which is also the length of the real images
-        z_dim: the dimension of the noise vector, a scalar
-        device: the device type
+
+        gen (nn.Module): the generator model, which returns an image
+            given z-dimensional noise
+        disc (nn.Module): the discriminator model, which returns a
+            single-dimensional prediction of real/fake
+        criterion (nn.Module): the loss function, which should be used
+            to compare the discriminator's predictions to the ground
+            truth reality of the images (e.g. fake = 0, real = 1)
+        real (Tensor): a batch of real images
+        num_images (int): the number of images the generator should
+            produce, which is also the length of the real images
+        z_dim (int): the dimension of the noise vector, a scalar
+        device (device): the device type
 
     Returns:
         Tensor: a torch scalar loss value for the current batch
+
     """
     noise_vector = get_noise(num_images, z_dim, device)
     generated_images = gen(noise_vector)
@@ -275,18 +287,22 @@ def get_gen_loss(gen, disc, criterion, num_images, z_dim, device):
     """Return the loss of the generator given inputs.
 
     Args:
-        gen: the generator model, which returns an image given z-dimensional noise
-        disc: the discriminator model, which returns a single-dimensional prediction of real/fake
-        criterion: the loss function, which should be used to compare
-               the discriminator's predictions to the ground truth reality of the images
-               (e.g. fake = 0, real = 1)
-        num_images: the number of images the generator should produce,
-                which is also the length of the real images
-        z_dim: the dimension of the noise vector, a scalar
-        device: the device type
+
+        gen (nn.Module): the generator model, which returns an image
+            given z-dimensional noise
+        disc (nn.Module): the discriminator model, which returns a
+            single-dimensional prediction of real/fake
+        criterion (nn.Module): the loss function, which should be used
+            to compare the discriminator's predictions to the ground
+            truth reality of the images (e.g. fake = 0, real = 1)
+        num_images (int): the number of images the generator should
+            produce, which is also the length of the real images
+        z_dim (int): the dimension of the noise vector, a scalar
+        device (device): the device type
 
     Returns:
         Tensor: a torch scalar loss value for the current batch
+
     """
     noise_vector = get_noise(num_images, z_dim, device)
     generated_images = gen(noise_vector)
@@ -302,15 +318,16 @@ def training_loop(
     """Training loop.
 
     Args:
-        gen: the generator model
-        disc: the discriminator model
-        gen_opt: analog model optimizer for the generator
-        disc_opt: analog model optimizer for the discriminator
-        criterion: criterion to compute loss
-        dataloader: Data set to train and evaluate the models
+        gen (nn.Module): the generator model
+        disc (nn.Module): the discriminator model
+        gen_opt (Optimizer): analog model optimizer for the generator
+        disc_opt (Optimizer): analog model optimizer for the discriminator
+        criterion (nn.Module): criterion to compute loss
+        dataloader (DataLoader): Data set to train and evaluate the models
         n_epochs (int): global parameter to define epochs number
         display_step (int): defines the period to display the training progress
     """
+    # pylint: disable=too-many-locals
     cur_step = 0
     mean_generator_loss = 0
     mean_discriminator_loss = 0
