+++ examples/17_resnet34_imagenet_conversion_to_analog.py
@@ -15,6 +15,7 @@
 resnet34 inspired network based on the paper:
 https://arxiv.org/abs/1512.03385
 """
+# pylint: disable=invalid-name
 
 # Imports from PyTorch.
 from torchvision.models import resnet34
